refactor: Move layout to layouts dir
